WIP: comparision report
